First 4 user stories and some extra bits.
As an air traffic controller
So I can get passengers to a destination
I want to instruct a plane to land at an airport

As an air traffic controller
So I can get passengers on the way to their destination
I want to instruct a plane to take off from an airport and confirm that it is no longer in the airport

As an air traffic controller
To ensure safety
I want to prevent landing when the airport is full

As the system designer
So that the software can be used for many different airports
I would like a default airport capacity that can be overridden as appropriate

Extra bits:
The airport will confirm when the plane has landed and taken off.
A plane can only land of take_off if it is a correct request.
The plane logs a timestamp whenever it lands or takes off.